test: isolate Gemini remainder regressions
